+++ src/mkvtoolnix-gui/main_window/preferences_dialog.cpp
@@ -29,6 +29,7 @@
 #include "mkvtoolnix-gui/util/side_by_side_multi_select.h"
 #include "mkvtoolnix-gui/util/string_list_configuration_widget.h"
 #include "mkvtoolnix-gui/util/widget.h"
+#include "qregularexpression.h"
 
 namespace mtx::gui {
 
@@ -122,6 +123,7 @@ PreferencesDialog::PreferencesDialog(QWidget *parent,
   ui->cbMAddBlurayCovers->setChecked(m_cfg.m_mergeAddBlurayCovers);
   ui->cbMAttachmentAlwaysSkipForExistingName->setChecked(m_cfg.m_mergeAttachmentsAlwaysSkipForExistingName);
 
+  setupForcedDisplayFlagSubtitles();
   setupFileColorsControls();
   setupProcessPriority();
   setupPlaylistScanningPolicy();
@@ -512,6 +514,14 @@ PreferencesDialog::setupToolTips() {
                    .arg(QYH("The language selected here is used for subtitle tracks for which their source file contains no such property and for which the language has not been derived from the file name."))
                    .arg(QYH("Depending on the setting below this language can also be used if the language in the source file is 'undetermined' ('und').")));
 
+  auto text = Q("<p>%1 %2</p>")
+    .arg(QYH("Subtitle files often contain the word 'forced' in their file name to signal that they're intended for 'forced display' only (e.g. when they speak Elfish in 'Lord of the Rings')."))
+    .arg(QYH("The GUI can set the 'forced display' flag for such tracks if the file name matches this regular expression."));
+
+  Util::setToolTip(ui->cbMForcedDisplayFlagSubtitles,   text);
+  Util::setToolTip(ui->leMForcedDisplayFlagSubtitlesRE, text);
+  Util::setToolTip(ui->pbMForcedDisplayFlagSubtitlesRERevert, QY("Revert the entry to its default value."));
+
   Util::setToolTip(ui->cbMDefaultSubtitleCharset, QY("If a character set is selected here, the program will automatically set the character set input to this value for newly added text subtitle tracks."));
 
   Util::setToolTip(ui->leMDefaultAdditionalCommandLineOptions, QY("The options entered here are set for all new multiplex jobs by default."));
@@ -625,6 +635,8 @@ PreferencesDialog::setupConnections() {
   connect(ui->cbMEnableMuxingTracksByLanguage,            &QCheckBox::toggled,                                           ui->tbMEnableMuxingTracksByLanguage,  &QLabel::setEnabled);
 
   connect(ui->pbMDeriveTrackLanguageRevertBoundaryChars,  &QPushButton::clicked,                                         this,                                 &PreferencesDialog::revertDeriveTrackLanguageFromFileNameChars);
+  connect(ui->pbMForcedDisplayFlagSubtitlesRERevert,      &QPushButton::clicked,                                         this,                                 &PreferencesDialog::revertForcedDisplayFlagSubtitlesRE);
+  connect(ui->cbMForcedDisplayFlagSubtitles,              &QCheckBox::toggled,                                           this,                                 &PreferencesDialog::enableForcedDisplayFlagSubtitlesControls);
 
   connect(ui->cbGuiRemoveJobs,                            &QCheckBox::toggled,                                           ui->cbGuiJobRemovalPolicy,            &QComboBox::setEnabled);
   connect(ui->cbGuiRemoveJobsOnExit,                      &QCheckBox::toggled,                                           ui->cbGuiJobRemovalOnExitPolicy,      &QComboBox::setEnabled);
@@ -937,6 +949,22 @@ PreferencesDialog::setupTabPositions() {
 }
 
 void
+PreferencesDialog::setupForcedDisplayFlagSubtitles() {
+  ui->cbMForcedDisplayFlagSubtitles->setChecked(m_cfg.m_deriveSubtitlesForcedFlagFromFileNames);
+  ui->leMForcedDisplayFlagSubtitlesRE->setText(m_cfg.m_regexForDerivingSubtitlesForcedFlagFromFileNames);
+
+  enableForcedDisplayFlagSubtitlesControls();
+}
+
+void
+PreferencesDialog::enableForcedDisplayFlagSubtitlesControls() {
+  auto enable = ui->cbMForcedDisplayFlagSubtitles->isChecked();
+
+  ui->leMForcedDisplayFlagSubtitlesRE->setEnabled(enable);
+  ui->pbMForcedDisplayFlagSubtitlesRERevert->setEnabled(enable);
+}
+
+void
 PreferencesDialog::setupDerivingTrackLanguagesFromFileName() {
   auto setupComboBox = [](QComboBox &cb, Util::Settings::DeriveLanguageFromFileNamePolicy policy) {
     cb.clear();
@@ -1173,6 +1201,8 @@ PreferencesDialog::save() {
   m_cfg.m_defaultVideoTrackLanguage                           = ui->ldwMDefaultVideoTrackLanguage->language();
   m_cfg.m_defaultSubtitleTrackLanguage                        = ui->ldwMDefaultSubtitleTrackLanguage->language();
   m_cfg.m_whenToSetDefaultLanguage                            = static_cast<Util::Settings::SetDefaultLanguagePolicy>(ui->cbMWhenToSetDefaultLanguage->currentData().toInt());
+  m_cfg.m_deriveSubtitlesForcedFlagFromFileNames              = ui->cbMForcedDisplayFlagSubtitles->isChecked();
+  m_cfg.m_regexForDerivingSubtitlesForcedFlagFromFileNames    = ui->leMForcedDisplayFlagSubtitlesRE->text();
   m_cfg.m_defaultSubtitleCharset                              = ui->cbMDefaultSubtitleCharset->currentData().toString();
   m_cfg.m_priority                                            = static_cast<Util::Settings::ProcessPriority>(ui->cbMProcessPriority->currentData().toInt());
   m_cfg.m_defaultAdditionalMergeOptions                       = ui->leMDefaultAdditionalCommandLineOptions->text();
@@ -1398,6 +1428,11 @@ PreferencesDialog::revertDeriveTrackLanguageFromFileNameChars() {
   ui->leMDeriveTrackLanguageBoundaryChars->setText(Util::Settings::defaultBoundaryCharsForDerivingLanguageFromFileName());
 }
 
+void
+PreferencesDialog::revertForcedDisplayFlagSubtitlesRE() {
+  ui->leMForcedDisplayFlagSubtitlesRE->setText(Util::Settings::defaultRegexForDerivingForcedDisplayFlagForSubtitlesFromFileName());
+}
+
 bool
 PreferencesDialog::verifyDeriveTrackLanguageSettings() {
   if (   (static_cast<Util::Settings::DeriveLanguageFromFileNamePolicy>(ui->cbMDeriveAudioTrackLanguageFromFileName   ->currentData().toInt()) == Util::Settings::DeriveLanguageFromFileNamePolicy::Never)
@@ -1430,6 +1465,22 @@ PreferencesDialog::verifyDeriveTrackLanguageSettings() {
 }
 
 bool
+PreferencesDialog::verifyDeriveForcedDisplayFlagSettings() {
+  auto reText = ui->leMForcedDisplayFlagSubtitlesRE->text();
+  QRegularExpression re{reText, QRegularExpression::CaseInsensitiveOption};
+
+  if (!reText.isEmpty() && re.isValid())
+    return true;
+
+  Util::MessageBox::critical(this)
+    ->title(QY("Invalid settings"))
+    .text(QY("The value for deriving the 'forced display' flag for subtitles from file names must be a valid regular expression."))
+    .exec();
+
+  return false;
+}
+
+bool
 PreferencesDialog::verifyRunProgramConfigurations() {
   for (auto tabIdx = 0, numTabs = ui->twJobsPrograms->count(); tabIdx < numTabs; ++tabIdx) {
     auto tab   = qobject_cast<PrefsRunProgramWidget *>(ui->twJobsPrograms->widget(tabIdx));
@@ -1459,6 +1510,7 @@ PreferencesDialog::verifyRunProgramConfigurations() {
 void
 PreferencesDialog::accept() {
   if (   verifyDeriveTrackLanguageSettings()
+      && verifyDeriveForcedDisplayFlagSettings()
       && verifyRunProgramConfigurations()) {
     rememberCurrentlySelectedPage();
     QDialog::accept();