+++ src/mkvtoolnix-gui/util/settings.cpp
@@ -718,6 +718,8 @@ Settings::loadDefaults(QSettings &reg) {
   m_defaultChapterLanguage             = mtx::bcp47::language_c::parse(to_utf8(reg.value(s_valDefaultChapterLanguage,       Q("und")).toString()));
   m_defaultSubtitleCharset             = reg.value(s_valDefaultSubtitleCharset).toString();
   m_defaultAdditionalMergeOptions      = reg.value(s_valDefaultAdditionalMergeOptions).toString();
+  m_deriveSubtitlesForcedFlagFromFileNames           = reg.value(s_valDefaultDeriveSubtitlesForcedFlagFromFileNames, true).toBool();
+  m_regexForDerivingSubtitlesForcedFlagFromFileNames = reg.value(s_valDefaultRegexForDerivingSubtitlesForcedFlagFromFileNames, defaultRegexForDerivingForcedDisplayFlagForSubtitlesFromFileName()).toString();
   reg.endGroup();               // defaults
 }
 
@@ -1076,6 +1078,8 @@ Settings::saveDefaults(QSettings &reg)
   reg.setValue(s_valDefaultChapterLanguage,        Q(m_defaultChapterLanguage.format()));
   reg.setValue(s_valDefaultSubtitleCharset,        m_defaultSubtitleCharset);
   reg.setValue(s_valDefaultAdditionalMergeOptions, m_defaultAdditionalMergeOptions);
+  reg.setValue(s_valDefaultDeriveSubtitlesForcedFlagFromFileNames,           m_deriveSubtitlesForcedFlagFromFileNames);
+  reg.setValue(s_valDefaultRegexForDerivingSubtitlesForcedFlagFromFileNames, m_regexForDerivingSubtitlesForcedFlagFromFileNames);
   reg.endGroup();               // defaults
 }
 
@@ -1373,6 +1377,11 @@ Settings::determineMediaInfoExePath() {
 QString
 Settings::defaultBoundaryCharsForDerivingLanguageFromFileName() {
   return Q("[](){}.+-=#");
+}
+
+QString
+Settings::defaultRegexForDerivingForcedDisplayFlagForSubtitlesFromFileName() {
+  return Q("[[\\](){}.+=#-]forced[[\\](){}.+=#-]");
 }
 
 QVector<QColor>