+++ src/mkvtoolnix-gui/util/settings.h
@@ -189,6 +189,8 @@ public:
   SetDefaultLanguagePolicy m_whenToSetDefaultLanguage;
   DeriveLanguageFromFileNamePolicy m_deriveAudioTrackLanguageFromFileNamePolicy, m_deriveVideoTrackLanguageFromFileNamePolicy, m_deriveSubtitleTrackLanguageFromFileNamePolicy;
   QString m_boundaryCharsForDerivingTrackLanguagesFromFileNames;
+  bool m_deriveSubtitlesForcedFlagFromFileNames;
+  QString m_regexForDerivingSubtitlesForcedFlagFromFileNames;
   QStringList m_recognizedTrackLanguagesInFileNames, m_mergePredefinedSplitSizes, m_mergePredefinedSplitDurations;
   QStringList m_mergePredefinedVideoTrackNames, m_mergePredefinedAudioTrackNames, m_mergePredefinedSubtitleTrackNames;
   QString m_chapterNameTemplate, m_ceTextFileCharacterSet, m_defaultSubtitleCharset, m_defaultAdditionalMergeOptions;
@@ -329,6 +331,7 @@ public:
   static QString prepareCacheDir(QString const &subDir);
 
   static QString defaultBoundaryCharsForDerivingLanguageFromFileName();
+  static QString defaultRegexForDerivingForcedDisplayFlagForSubtitlesFromFileName();
 
   static QVector<QColor> defaultFileColors();